fix: resolve residual quality findings (#1206)
